[scheduler] Refactor edit dialog General tab into section components
Extract the date/time, resource & color, and description sections of the
edit dialog General tab into self-contained components, so GeneralTab is
just the default JSX composition of the three (#22867). No public or
/internals API changes; the export flip happens with the General-tab
slot (#22871).
