+++ packages/x-scheduler/src/internals/components/event-dialog/EventDialog.test.tsx
@@ -12,6 +12,7 @@ import { SchedulerStoreContext } from '@mui/x-scheduler-internals/use-scheduler-
 import type { SchedulerEvent } from '@mui/x-scheduler/models';
 import { EventDialogContent } from './EventDialog';
 import { EventCalendarProvider } from '../EventCalendarProvider';
+import { eventCalendarClasses } from '../../../event-calendar/eventCalendarClasses';
 
 const personalResource = ResourceBuilder.new().title('Personal').eventColor('teal').build();
 
@@ -42,6 +43,23 @@ describe('<EventDialogContent /> — community (no recurring-events plugin)', ()
   };
 
   const { render } = createSchedulerRenderer();
+
+  it('should render the general tab sections in the default order', () => {
+    render(
+      <EventCalendarProvider events={[DEFAULT_EVENT]} resources={resources}>
+        <EventDialogContent open {...defaultProps} />
+      </EventCalendarProvider>,
+    );
+
+    const legends = Array.from(
+      document.body.getElementsByClassName(eventCalendarClasses.eventDialogSectionHeaderTitle),
+    );
+    expect(legends.map((legend) => legend.textContent)).to.deep.equal([
+      'Date & time',
+      'Resource & color',
+    ]);
+    expect(screen.getByRole('textbox', { name: 'Description' })).not.to.equal(null);
+  });
 
   it('should not render the recurrence tab when no slot is provided', () => {
     render(
